+++ spec/whois/record/parser/responses/whois.register.com/status_registered_spec.rb
@@ -23,7 +23,7 @@ describe Whois::Record::Parser::WhoisRegisterCom, "status_registered.expected" d
 
   describe "#status" do
     it do
-      @parser.status.should == :registered
+      lambda { @parser.status }.should raise_error(Whois::PropertyNotSupported)
     end
   end
   describe "#available?" do
@@ -44,7 +44,7 @@ describe Whois::Record::Parser::WhoisRegisterCom, "status_registered.expected" d
   end
   describe "#updated_on" do
     it do
-      @parser.updated_on.should == nil
+      lambda { @parser.updated_on }.should raise_error(Whois::PropertyNotSupported)
     end
   end
   describe "#expires_on" do
@@ -58,8 +58,7 @@ describe Whois::Record::Parser::WhoisRegisterCom, "status_registered.expected" d
       @parser.registrar.should be_a(Whois::Record::Registrar)
       @parser.registrar.id.should           == nil
       @parser.registrar.name.should         == "Register.com"
-      @parser.registrar.organization.should == "Register.com"
-      @parser.registrar.url.should          == "http://register.com"
+      @parser.registrar.url.should          == "http://www.register.com/"
     end
   end
   describe "#registrant_contacts" do
@@ -68,13 +67,12 @@ describe Whois::Record::Parser::WhoisRegisterCom, "status_registered.expected" d
       @parser.registrant_contacts.should have(1).items
       @parser.registrant_contacts[0].should be_a(Whois::Record::Contact)
       @parser.registrant_contacts[0].type.should         == Whois::Record::Contact::TYPE_REGISTRANT
-      @parser.registrant_contacts[0].id.should           == nil
       @parser.registrant_contacts[0].name.should         == "Domain Registrar"
       @parser.registrant_contacts[0].organization.should == "Register.Com, Inc."
       @parser.registrant_contacts[0].address.should      == "575 8th Avenue"
       @parser.registrant_contacts[0].city.should         == "New York"
-      @parser.registrant_contacts[0].state.should        == "NY"
       @parser.registrant_contacts[0].zip.should          == "10018"
+      @parser.registrant_contacts[0].state.should        == "NY"
       @parser.registrant_contacts[0].country_code.should == "US"
       @parser.registrant_contacts[0].phone.should        == "+1.9027492701"
       @parser.registrant_contacts[0].fax.should          == nil
@@ -87,13 +85,12 @@ describe Whois::Record::Parser::WhoisRegisterCom, "status_registered.expected" d
       @parser.admin_contacts.should have(1).items
       @parser.admin_contacts[0].should be_a(Whois::Record::Contact)
       @parser.admin_contacts[0].type.should         == Whois::Record::Contact::TYPE_ADMIN
-      @parser.admin_contacts[0].id.should           == nil
       @parser.admin_contacts[0].name.should         == "Domain Registrar"
       @parser.admin_contacts[0].organization.should == "Register.Com, Inc."
       @parser.admin_contacts[0].address.should      == "575 8th Avenue"
       @parser.admin_contacts[0].city.should         == "New York"
-      @parser.admin_contacts[0].state.should        == "NY"
       @parser.admin_contacts[0].zip.should          == "10018"
+      @parser.admin_contacts[0].state.should        == "NY"
       @parser.admin_contacts[0].country_code.should == "US"
       @parser.admin_contacts[0].phone.should        == "+1.9027492701"
       @parser.admin_contacts[0].fax.should          == nil
@@ -106,13 +103,12 @@ describe Whois::Record::Parser::WhoisRegisterCom, "status_registered.expected" d
       @parser.technical_contacts.should have(1).items
       @parser.technical_contacts[0].should be_a(Whois::Record::Contact)
       @parser.technical_contacts[0].type.should         == Whois::Record::Contact::TYPE_TECHNICAL
-      @parser.technical_contacts[0].id.should           == nil
       @parser.technical_contacts[0].name.should         == "Domain Registrar"
       @parser.technical_contacts[0].organization.should == "Register.Com, Inc."
       @parser.technical_contacts[0].address.should      == "575 8th Avenue"
       @parser.technical_contacts[0].city.should         == "New York"
-      @parser.technical_contacts[0].state.should        == "NY"
       @parser.technical_contacts[0].zip.should          == "10018"
+      @parser.technical_contacts[0].state.should        == "NY"
       @parser.technical_contacts[0].country_code.should == "US"
       @parser.technical_contacts[0].phone.should        == "+1.9027492701"
       @parser.technical_contacts[0].fax.should          == nil
@@ -125,22 +121,16 @@ describe Whois::Record::Parser::WhoisRegisterCom, "status_registered.expected" d
       @parser.nameservers.should have(6).items
       @parser.nameservers[0].should be_a(Whois::Record::Nameserver)
       @parser.nameservers[0].name.should == "ns2.register.com"
-      @parser.nameservers[0].ipv4.should == nil
       @parser.nameservers[1].should be_a(Whois::Record::Nameserver)
       @parser.nameservers[1].name.should == "ns3.register.com"
-      @parser.nameservers[1].ipv4.should == nil
       @parser.nameservers[2].should be_a(Whois::Record::Nameserver)
       @parser.nameservers[2].name.should == "ns4.register.com"
-      @parser.nameservers[2].ipv4.should == nil
       @parser.nameservers[3].should be_a(Whois::Record::Nameserver)
       @parser.nameservers[3].name.should == "ns5.register.com"
-      @parser.nameservers[3].ipv4.should == nil
       @parser.nameservers[4].should be_a(Whois::Record::Nameserver)
       @parser.nameservers[4].name.should == "ns6.register.com"
-      @parser.nameservers[4].ipv4.should == nil
       @parser.nameservers[5].should be_a(Whois::Record::Nameserver)
       @parser.nameservers[5].name.should == "ns1.register.com"
-      @parser.nameservers[5].ipv4.should == nil
     end
   end
 end